+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: Main
+sample.Main-Class: sample.Main
