+++ src/main/java/gui/DialogBox.java
@@ -1,3 +1,5 @@
+package gui;
+
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
 import javafx.scene.image.ImageView;
