+++ src/main/java/gui/Launcher.java
@@ -1,3 +1,5 @@
+package gui;
+
 import javafx.application.Application;
 
 /**
